+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Fix: wrap: false returning inconsistent data types (@CacheControl)
 - Fix: Ensure throwing with a bad result type
+- Fix: Avoid erring when value before parent selector is falsey
 - Enhancement: Allow path as array in non-object signature
 - Linting (ESLint): As per latest ash-nazg
 - Linting (ESLint): Remove redundant "use strict" with switch to ESM